+++ src/app/projects/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next';
 
 import phrases from 'data/phrases';
-import projectData from 'data/projectData';
+import projectData from 'data/project-data';
 import siteMetadata from 'data/siteMetadata.mjs';
 
 import ProjectCard from '@/components/card-and-list/project-card';
